+++ packages/core/useDark/index.md
@@ -85,4 +85,3 @@ const isDark = useDark({
 - `useColorMode`
 - `usePreferredDark`
 - `useStorage`
-
